+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - allow omitting space between `//` and `coverage` in coverage ignore comments
 - allow text after coverage ignore comments
-- handle unbalanced ignore comments instead of silently erroring
+- throw FormatException when encountering unbalanced ignore comments instead of silently erroring
 
 ## 1.6.3